Remove EclProblemTpfa, folding all methods into EclProblem.
